refactor(sybra): extract completion pipeline
Moves AgentCompletionHandler, DomainHandler, and the fix-review/
human-review completion routing out of internal/sybra into
internal/sybra/completion (nested per the god-package extraction
convention). Handler fields stay private with a Config-based New()
constructor; the human-review callback and conflict-recovery callback
are wired as plain func values from internal/sybra to avoid an import
cycle back into the package that constructs the handler.
