+++ internal/sybra/app_init.go
@@ -699,7 +699,7 @@ func (a *App) initWorkflowEngine() {
 		a.agentOrch.SetConflictRecovery(a.reviewer.RecoverStaleBranchConflict)
 	}
 	// Workflow completion moves to wireServices so the callback closure binds
-	// to the AgentCompletionHandler constructed there.
+	// to the completion.Handler constructed there.
 }
 
 func (a *App) initAgentConfig() {
