+++ internal/workflow/engine_steps_resume.go
@@ -11,7 +11,7 @@ import (
 // resumeStatusVar is always set (to the task's status at the moment recovery
 // began) so a task that had no active workflow still gets its visible status
 // restored before the normal status-driven cascade (see
-// AgentCompletionHandler.OnWorkflowComplete) picks up wherever it left off.
+// completion.Handler.OnWorkflowComplete) picks up wherever it left off.
 const (
 	resumeWorkflowIDVar   = "resume_workflow_id"
 	resumeWorkflowStepVar = "resume_workflow_step"
@@ -29,7 +29,7 @@ const (
 // When no workflow was captured (the task had no active workflow at the
 // time recovery began), this is a no-op beyond restoring status: the step
 // completes normally, the recovery workflow ends, and
-// AgentCompletionHandler.OnWorkflowComplete's existing status-driven cascade
+// completion.Handler.OnWorkflowComplete's existing status-driven cascade
 // dispatches whatever workflow matches the now-restored task.status — the
 // same mechanism every other builtin workflow chain (plan → implement →
 // review → ...) already relies on.
